feat: transformed timezones array into a mixin
